+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # 0.18.0
 
+## Link (React Router)
+
+Our Link component now supports the React Router. Instead of passing a `href` prop, pass a `to` prop and it will use React Router to navigate.
+
 ## Pod Updates
 
-* Pod can now receive a prop of `onEdit` - if this is a String it will use it as a `href` link, if it is a Function it will apply it as an `onClick` handler.
+* Pod can now receive a prop of `onEdit` - if this is a String it will use it as a `to` prop on a Link component, if it is a Function it will apply it as an `onClick` handler, if it is an object it will apply it's props to the Link.
 * Pod has an additional padding size added of `extra-large`.
 
 # 0.17.1